Move prometheus hacks into its own file
